requirements.txt
packaging
tqdm
scikit-learn
PyYAML
easydict
torch_tensorrt
jupyterlab
pycuda